feat: add BasmalaHeader and UnifiedChunkAyat components with animations for Ayah action buttons
