+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/quran/surah/component/AyatOfSurah.kt
@@ -116,7 +116,7 @@ private fun rememberPreRenderedChunks(ayahChunks: List<List<Ayah>>): List<Annota
     }
 }
 
-@Preview(showBackground = true)
+@Preview()
 @Composable
 fun AyatOfSurahPreview() {
     QuranTheme {
