+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/quran/surah/component/ScrollTracking.kt
@@ -98,8 +98,7 @@ fun HideAyahActionButtonsOnScroll(
     LaunchedEffect(lazyListState, state.isAyahActionButtonsVisible) {
         snapshotFlow { lazyListState.isScrollInProgress }
             .collect { isScrolling ->
-                if (isScrolling && state.isAyahActionButtonsVisible)
-                    listener.onDismissActionButtons()
+                if (isScrolling && state.isAyahActionButtonsVisible) listener.onDismissActionButtons()
             }
     }
 }
